Move modularized components to their own files
